Add support for the Cisco WLC
I hoped that somebody would beat me to add the Cisco WLC upstream...
Though there was a lot of interest in testing such a patch so here it
is. Especially testing on versions 4.x is needed as I could test it
only on 5.2.x and 6.0.x.

The attached patch is heavily based on the scripts posted by Ryan West
here on rancid-discuss
(http://www.shrubbery.net/pipermail/rancid-discuss/2010-February/004652.html).

The patch applies on top of 2.3.3 and the license of the portions done
by me is as always "whatever it takes to get the patch accepted upstream".

My modifications:
-----------------
- Rename ciscowlc5 to wlcrancid.
- Rename wlogin to wlclogin.
- Merge support for version 4.x into wlcrancid (*untested*).
- Add -V command line arg to wlclogin/wlcrancid.
- Remove the "!--WLC Begin Config Data--" markers.
- Cleanups in wlcrancid (indentation, whitespace).
- Added the needed configure/Makefile changes.

Todo:
-----
I might work on this items, when I get the time for it is a totally
different story.
- Add man pages for wlclogin/wlcrancid
- From the parsed info generate the headers (comments) that rancid adds,
e.g.:
+ !Chassis type:
+ !Processor ID:
+ !Image: Software:
- Get rid of the "!WLC Show Udi Start" and "!WLC Show Sysinfo Start"
markers too.